FULL BLUEALLIANCE CSV SUPPORT WOOT
Just frickign copy the entire match schedule table into Excel or
OpenOffice from the BlueAlliance site, then save it as a CSV with
default settings. The program can read it! And load an entire event with
the match data! Plus the team numbers! Heck yeah! :D

It doesn't load team names, but shhhh. It's awesome anyway.